+++ packages/lib/src/model/Model.ts
@@ -1,5 +1,6 @@
 import { applySet } from "../action/applySet"
-import { getCurrentActionContext } from "../action/context"
+import { ActionContextActionType, getCurrentActionContext } from "../action/context"
+import { wrapInAction } from "../action/wrapInAction"
 import { getGlobalConfig } from "../globalConfig"
 import { memoTransformCache } from "../propTransform/propTransform"
 import { typesObject } from "../typeChecking/object"
@@ -7,8 +8,9 @@ import { typesString } from "../typeChecking/primitives"
 import { tProp } from "../typeChecking/tProp"
 import { LateTypeChecker } from "../typeChecking/TypeChecker"
 import { typesUnchecked } from "../typeChecking/unchecked"
-import { assertIsObject, failure } from "../utils"
+import { assertIsObject, failure, propNameToSetterActionName } from "../utils"
 import {
+  AbstractModelClass,
   AnyModel,
   BaseModel,
   baseModelPropNames,
@@ -23,13 +25,14 @@ import { ModelConstructorOptions } from "./ModelConstructorOptions"
 import { getInternalModelClassPropsInfo, setInternalModelClassPropsInfo } from "./modelPropsInfo"
 import { modelMetadataSymbol, modelUnwrappedClassSymbol } from "./modelSymbols"
 import {
+  AnyModelProp,
   idProp,
-  ModelProp,
   ModelProps,
   ModelPropsToInstanceCreationData,
   ModelPropsToInstanceData,
   ModelPropsToPropsCreationData,
   ModelPropsToPropsData,
+  ModelPropsToSetterActions,
   prop,
 } from "./prop"
 import { assertIsModelClass } from "./utils"
@@ -66,7 +69,8 @@ export interface _Model<SuperModel, TProps extends ModelProps> {
       this[typeof instanceCreationDataSymbol],
       ExtractModelIdProp<TProps> & string
     > &
-    Omit<this[typeof instanceDataSymbol], keyof AnyModel>
+    Omit<this[typeof instanceDataSymbol], keyof AnyModel> &
+    ModelPropsToSetterActions<TProps>
 }
 
 /**
@@ -90,21 +94,17 @@ export type ExtractModelIdProp<TProps extends ModelProps> = {
  * Base abstract class for models that extends another model.
  *
  * @typeparam TProps New model properties type.
- * @typeparam TBaseModelClass Base class type.
+ * @typeparam TModel Model type.
  * @param baseModel Base model type.
  * @param modelProps Model properties.
  * @returns
  */
-export function ExtendedModel<TProps extends ModelProps, TBaseModelClass>(
-  baseModel: TBaseModelClass,
+export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel>(
+  baseModel: AbstractModelClass<TModel>,
   modelProps: TProps
-): _Model<
-  TBaseModelClass & Object extends ModelClass<infer M> ? M : never,
-  AddModelIdPropIfNeeded<TProps>
-> {
+): _Model<TModel, AddModelIdPropIfNeeded<TProps>> {
   assertIsModelClass(baseModel, "baseModel")
 
-  // note that & Object is there to support abstract classes
   return internalModel(modelProps, baseModel as any)
 }
 
@@ -247,6 +247,20 @@ function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
 
   Object.defineProperties(CustomBaseModel.prototype, extraDescriptors)
 
+  // add setter actions to prototype
+  for (const [propName, propData] of Object.entries(modelProps)) {
+    if (propData.options.setterAction) {
+      const setterActionName = propNameToSetterActionName(propName)
+      CustomBaseModel.prototype[setterActionName] = wrapInAction({
+        name: setterActionName,
+        fn: function (this: any, value: any) {
+          this[propName] = value
+        },
+        actionType: ActionContextActionType.Sync,
+      })
+    }
+  }
+
   return CustomBaseModel
 }
 
@@ -258,7 +272,7 @@ function _inheritsLoose(subClass: any, superClass: any) {
 
 function createModelPropDescriptor(
   modelPropName: string,
-  modelProp: ModelProp<any, any, any> | undefined,
+  modelProp: AnyModelProp | undefined,
   enumerable: boolean
 ): PropertyDescriptor {
   return {
@@ -281,7 +295,7 @@ function createModelPropDescriptor(
 
 function getModelInstanceDataField<M extends AnyModel>(
   model: M,
-  modelProp: ModelProp<any, any, any> | undefined,
+  modelProp: AnyModelProp | undefined,
   modelPropName: keyof ModelInstanceData<M>
 ): ModelInstanceData<M>[typeof modelPropName] {
   const transform = modelProp ? modelProp.transform : undefined
@@ -302,11 +316,11 @@ function getModelInstanceDataField<M extends AnyModel>(
 
 function setModelInstanceDataField<M extends AnyModel>(
   model: M,
-  modelProp: ModelProp<any, any, any> | undefined,
+  modelProp: AnyModelProp | undefined,
   modelPropName: keyof ModelInstanceData<M>,
   value: ModelInstanceData<M>[typeof modelPropName]
 ): void {
-  if (modelProp?.options.setterAction && !getCurrentActionContext()) {
+  if (modelProp?.options.setterAction === "assign" && !getCurrentActionContext()) {
     // use apply set instead to wrap it in an action
     applySet(model, modelPropName as any, value)
     return