+++ packages/lib/src/model/prop.ts
@@ -9,9 +9,20 @@ import { IsOptionalValue } from "../utils/types"
  */
 export interface ModelPropOptions {
   /**
-   * Set to `true` to automatically wrap the property setter in a model action (defaults to `false`).
+   * Set to `true` to automatically generate a property setter in a model action (defaults to `false`).
+   * Set to `assign` to get the old behaviour of making the property assignable.
    */
-  readonly setterAction?: boolean
+  readonly setterAction?: boolean | "assign"
+}
+
+/**
+ * Model property options with setterAction set to true.
+ */
+export interface ModelPropOptionsWithSetterAction extends ModelPropOptions {
+  /**
+   * Set to `true` to automatically generate a property setter in a model action (defaults to `false`).
+   */
+  readonly setterAction: true
 }
 
 /**
@@ -28,7 +39,8 @@ export interface ModelProp<
   TIsOptional,
   TInstanceValue = TPropValue,
   TInstanceCreationValue = TPropCreationValue,
-  TIsId extends boolean = false
+  TIsId extends boolean = false,
+  THasSetterAction = never
 > {
   $propValueType: TPropValue
   $propCreationValueType: TPropCreationValue
@@ -36,6 +48,7 @@ export interface ModelProp<
   $instanceCreationValueType: TInstanceCreationValue
   $isOptional: TIsOptional
   $isId: TIsId
+  $hasSetterAction: THasSetterAction
 
   defaultFn: (() => TPropValue) | typeof noDefaultValue
   defaultValue: TPropValue | typeof noDefaultValue
@@ -45,10 +58,15 @@ export interface ModelProp<
 }
 
 /**
+ * Any model property.
+ */
+export type AnyModelProp = ModelProp<any, any, any, any, any, any, any>
+
+/**
  * Model properties.
  */
 export interface ModelProps {
-  [k: string]: ModelProp<any, any, any, any, any, any>
+  [k: string]: AnyModelProp
 }
 
 export type OptionalModelProps<MP extends ModelProps> = {
@@ -76,6 +94,12 @@ export type ModelPropsToInstanceCreationData<MP extends ModelProps> = O.Optional
   },
   OptionalModelProps<MP>
 >
+
+export type ModelPropsToSetterActions<MP extends ModelProps> = {
+  [k in keyof MP as MP[k]["$hasSetterAction"] & `set${Capitalize<k & string>}`]: (
+    value: MP[k]["$instanceValueType"]
+  ) => void
+}
 
 /**
  * A property that will be used as model id, replacing $modelId.
@@ -107,6 +131,22 @@ export type MaybeOptionalModelProp<TPropValue, TInstanceValue = TPropValue> = Mo
 >
 
 /**
+ * A model prop that maybe / maybe not is optional, depending on if the value can take undefined, with a setter action.
+ */
+export type MaybeOptionalModelPropWithSetterAction<
+  TPropValue,
+  TInstanceValue = TPropValue
+> = ModelProp<
+  TPropValue,
+  TPropValue,
+  IsOptionalValue<TPropValue, string, never>,
+  TInstanceValue,
+  TInstanceValue,
+  false,
+  string
+>
+
+/**
  * A model prop that is definitely optional.
  */
 export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelProp<
@@ -116,6 +156,39 @@ export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelPr
   TInstanceValue,
   TInstanceValue | null | undefined
 >
+
+/**
+ * A model prop that is definitely optional, with a setter action.
+ */
+export type OptionalModelPropWithSetterAction<TPropValue, TInstanceValue = TPropValue> = ModelProp<
+  TPropValue,
+  TPropValue | null | undefined,
+  string,
+  TInstanceValue,
+  TInstanceValue | null | undefined,
+  false,
+  string
+>
+
+/**
+ * Defines a model property, with an optional function to generate a default value
+ * if the input snapshot / model creation data is `null` or `undefined`.
+ *
+ * Example:
+ * ```ts
+ * x: prop(() => 10) // an optional number, with a default value of 10
+ * x: prop<number[]>(() => []) // an optional number array, with a default empty array
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param defaultFn Default value generator function.
+ * @param options Model property options.
+ * @returns
+ */
+export function prop<TValue>(
+  defaultFn: () => TValue,
+  options: ModelPropOptionsWithSetterAction
+): OptionalModelPropWithSetterAction<TValue>
 
 /**
  * Defines a model property, with an optional function to generate a default value
@@ -155,6 +228,27 @@ export function prop<TValue>(
  */
 export function prop<TValue>(
   defaultValue: OnlyPrimitives<TValue>,
+  options: ModelPropOptionsWithSetterAction
+): OptionalModelPropWithSetterAction<TValue>
+
+/**
+ * Defines a model property, with an optional default value
+ * if the input snapshot / model creation data is `null` or `undefined`.
+ * You should only use this with primitive values and never with object values
+ * (array, model, object, etc).
+ *
+ * Example:
+ * ```ts
+ * x: prop(10) // an optional number, with a default value of 10
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param defaultValue Default primitive value.
+ * @param options Model property options.
+ * @returns
+ */
+export function prop<TValue>(
+  defaultValue: OnlyPrimitives<TValue>,
   options?: ModelPropOptions
 ): OptionalModelProp<TValue>
 
@@ -171,9 +265,30 @@ export function prop<TValue>(
  * @param options Model property options.
  * @returns
  */
+export function prop<TValue>(
+  options: ModelPropOptionsWithSetterAction
+): MaybeOptionalModelPropWithSetterAction<TValue>
+
+/**
+ * Defines a model property with no default value.
+ *
+ * Example:
+ * ```ts
+ * x: prop<number>() // a required number
+ * x: prop<number | undefined>() // an optional number, which defaults to undefined
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param options Model property options.
+ * @returns
+ */
 export function prop<TValue>(options?: ModelPropOptions): MaybeOptionalModelProp<TValue>
 
-export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any> {
+// base
+export function prop<TValue>(
+  arg1?: any,
+  arg2?: any
+): ModelProp<TValue, any, any, TValue, any, any, any> {
   let def: any
   let opts: ModelPropOptions = {}
   let hasDefaultValue = false
@@ -204,6 +319,7 @@ export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any
     $instanceValueType: null as any,
     $instanceCreationValueType: null as any,
     $isId: null as never,
+    $hasSetterAction: null as any,
 
     defaultFn: hasDefaultValue && isDefFn ? def : noDefaultValue,
     defaultValue: hasDefaultValue && !isDefFn ? def : noDefaultValue,