+++ packages/lib/src/propTransform/propTransform.ts
@@ -1,5 +1,5 @@
 import { AnyModel } from "../model/BaseModel"
-import { ModelProp, noDefaultValue } from "../model/prop"
+import { AnyModelProp, noDefaultValue } from "../model/prop"
 import { addLateInitializationFunction, failure } from "../utils"
 
 /**
@@ -178,10 +178,10 @@ function toMemoPropTransform<TProp, TData>(
  * @internal
  */
 export function transformedProp(
-  prop: ModelProp<any, any, any, any, any>,
+  prop: AnyModelProp,
   transform: PropTransform<any, any>,
   transformDefault: boolean
-): ModelProp<any, any, any, any, any> {
+): AnyModelProp {
   if (prop.transform) {
     throw failure("a property cannot have more than one transform")
   }