+++ packages/lib/test/model/setterAction.test.ts
@@ -1,20 +1,33 @@
 import { computed } from "mobx"
-import { addActionMiddleware, model, Model, modelAction, prop } from "../../src"
+import { addActionMiddleware, model, Model, prop, tProp } from "../../src"
 import "../commonSetup"
 import { autoDispose } from "../utils"
 
 @model("P")
 export class P extends Model({
   y: prop(0, { setterAction: true }),
+  z: tProp(0, { setterAction: true }),
+  sy: prop(0, { setterAction: "assign" }),
+  sz: tProp(0, { setterAction: "assign" }),
 }) {
   @computed
-  get cv() {
+  get cy() {
     return this.y * 10
   }
 
-  @modelAction
-  setY(n: number) {
-    this.y = n
+  @computed
+  get cz() {
+    return this.z * 10
+  }
+
+  @computed
+  get csy() {
+    return this.sy * 10
+  }
+
+  @computed
+  get csz() {
+    return this.sz * 10
   }
 }
 
@@ -43,18 +56,17 @@ test("setterAction", () => {
   )
   expect(events.length).toBe(0)
 
-  p.y = 5
+  p.setY(5)
   expect(p.y).toBe(5)
   expect(p.$.y).toBe(5)
-  expect(p.cv).toBe(50)
+  expect(p.cy).toBe(50)
 
   expect(events).toMatchInlineSnapshot(`
     Array [
       Object {
         "ctx": Object {
-          "actionName": "$$applySet",
+          "actionName": "setY",
           "args": Array [
-            "y",
             5,
           ],
           "data": Object {},
@@ -63,7 +75,10 @@ test("setterAction", () => {
           "target": P {
             "$": Object {
               "$modelId": "id-1",
+              "sy": 0,
+              "sz": 0,
               "y": 5,
+              "z": 0,
             },
             "$modelType": "P",
           },
@@ -73,9 +88,8 @@ test("setterAction", () => {
       },
       Object {
         "ctx": Object {
-          "actionName": "$$applySet",
+          "actionName": "setY",
           "args": Array [
-            "y",
             5,
           ],
           "data": Object {},
@@ -84,7 +98,10 @@ test("setterAction", () => {
           "target": P {
             "$": Object {
               "$modelId": "id-1",
+              "sy": 0,
+              "sz": 0,
               "y": 5,
+              "z": 0,
             },
             "$modelType": "P",
           },
@@ -97,18 +114,18 @@ test("setterAction", () => {
   `)
   events.length = 0
 
-  // check that it is not wrapped in action when used inside an action already
-  p.setY(10)
-  expect(p.y).toBe(10)
-  expect(p.$.y).toBe(10)
-  expect(p.cv).toBe(100)
+  p.setZ(5)
+  expect(p.z).toBe(5)
+  expect(p.$.z).toBe(5)
+  expect(p.cz).toBe(50)
+
   expect(events).toMatchInlineSnapshot(`
     Array [
       Object {
         "ctx": Object {
-          "actionName": "setY",
+          "actionName": "setZ",
           "args": Array [
-            10,
+            5,
           ],
           "data": Object {},
           "parentContext": undefined,
@@ -116,7 +133,10 @@ test("setterAction", () => {
           "target": P {
             "$": Object {
               "$modelId": "id-1",
-              "y": 10,
+              "sy": 0,
+              "sz": 0,
+              "y": 5,
+              "z": 5,
             },
             "$modelType": "P",
           },
@@ -126,9 +146,9 @@ test("setterAction", () => {
       },
       Object {
         "ctx": Object {
-          "actionName": "setY",
+          "actionName": "setZ",
           "args": Array [
-            10,
+            5,
           ],
           "data": Object {},
           "parentContext": undefined,
@@ -136,7 +156,130 @@ test("setterAction", () => {
           "target": P {
             "$": Object {
               "$modelId": "id-1",
-              "y": 10,
+              "sy": 0,
+              "sz": 0,
+              "y": 5,
+              "z": 5,
+            },
+            "$modelType": "P",
+          },
+          "type": "sync",
+        },
+        "event": "action finished",
+        "result": undefined,
+      },
+    ]
+  `)
+  events.length = 0
+
+  p.sy = 5
+  expect(p.sy).toBe(5)
+  expect(p.$.sy).toBe(5)
+  expect(p.csy).toBe(50)
+
+  expect(events).toMatchInlineSnapshot(`
+    Array [
+      Object {
+        "ctx": Object {
+          "actionName": "$$applySet",
+          "args": Array [
+            "sy",
+            5,
+          ],
+          "data": Object {},
+          "parentContext": undefined,
+          "rootContext": [Circular],
+          "target": P {
+            "$": Object {
+              "$modelId": "id-1",
+              "sy": 5,
+              "sz": 0,
+              "y": 5,
+              "z": 5,
+            },
+            "$modelType": "P",
+          },
+          "type": "sync",
+        },
+        "event": "action started",
+      },
+      Object {
+        "ctx": Object {
+          "actionName": "$$applySet",
+          "args": Array [
+            "sy",
+            5,
+          ],
+          "data": Object {},
+          "parentContext": undefined,
+          "rootContext": [Circular],
+          "target": P {
+            "$": Object {
+              "$modelId": "id-1",
+              "sy": 5,
+              "sz": 0,
+              "y": 5,
+              "z": 5,
+            },
+            "$modelType": "P",
+          },
+          "type": "sync",
+        },
+        "event": "action finished",
+        "result": undefined,
+      },
+    ]
+  `)
+  events.length = 0
+
+  p.sz = 5
+  expect(p.sz).toBe(5)
+  expect(p.$.sz).toBe(5)
+  expect(p.csz).toBe(50)
+
+  expect(events).toMatchInlineSnapshot(`
+    Array [
+      Object {
+        "ctx": Object {
+          "actionName": "$$applySet",
+          "args": Array [
+            "sz",
+            5,
+          ],
+          "data": Object {},
+          "parentContext": undefined,
+          "rootContext": [Circular],
+          "target": P {
+            "$": Object {
+              "$modelId": "id-1",
+              "sy": 5,
+              "sz": 5,
+              "y": 5,
+              "z": 5,
+            },
+            "$modelType": "P",
+          },
+          "type": "sync",
+        },
+        "event": "action started",
+      },
+      Object {
+        "ctx": Object {
+          "actionName": "$$applySet",
+          "args": Array [
+            "sz",
+            5,
+          ],
+          "data": Object {},
+          "parentContext": undefined,
+          "rootContext": [Circular],
+          "target": P {
+            "$": Object {
+              "$modelId": "id-1",
+              "sy": 5,
+              "sz": 5,
+              "y": 5,
+              "z": 5,
             },
             "$modelType": "P",
           },